+++ src/main/java/net/soulsweaponry/entity/projectile/WitheredWabbajackProjectile.java
@@ -24,16 +24,20 @@ import net.minecraft.world.GameRules;
 import net.minecraft.world.ServerWorldAccess;
 import net.minecraft.world.World;
 import net.minecraft.world.event.GameEvent;
-import net.soulsweaponry.items.staff.WitheredWabbajack.LuckType;
 import net.soulsweaponry.registry.EntityRegistry;
 import net.soulsweaponry.registry.SoundRegistry;
 import net.soulsweaponry.particles.ParticleHandler;
+import net.soulsweaponry.util.LuckChosenObject;
+import net.soulsweaponry.util.WeaponUtil;
 
 import java.util.ArrayList;
-import java.util.HashMap;
+import java.util.List;
 import java.util.Random;
 
 public class WitheredWabbajackProjectile extends WitherSkullEntity {
+
+    private static final List<LuckChosenObject<EntityType<?>>> ENTITIES = new ArrayList<>();
+    private static final List<LuckChosenObject<CollisionEffect>> COLLISIONS = new ArrayList<>();
 
     public WitheredWabbajackProjectile(EntityType<? extends WitheredWabbajackProjectile> entityType, World world) {
         super(entityType, world);
@@ -70,16 +74,16 @@ public class WitheredWabbajackProjectile extends WitherSkullEntity {
             if (entity instanceof LivingEntity target && owner instanceof LivingEntity user) {
                 Random random = new Random();
                 int rng = random.nextInt(10) + 1;
-                int power = this.getBound(75 , 5, user) + this.getLuckFactor(user) * 5;
-                int amplifier = this.getBound(3 , 1, user) + this.getLuckFactor(user)/2;
-                int duration = this.getBound(300 , 50, user) + this.getLuckFactor(user) * 50;
+                int power = this.getBound(75 , 5, user) + WeaponUtil.getLuckFactor(user) * 5;
+                int amplifier = this.getBound(3 , 1, user) + WeaponUtil.getLuckFactor(user)/2;
+                int duration = this.getBound(300 , 50, user) + WeaponUtil.getLuckFactor(user) * 50;
                 switch (rng) {
                     case 1, 2, 3 ->
                             target.addStatusEffect(new StatusEffectInstance(this.getRandomEffect(true), duration, amplifier));
                     case 4, 5 ->
                             user.addStatusEffect(new StatusEffectInstance(this.getRandomEffect(false), duration, amplifier));
                     case 6 -> {
-                        boolean luck = this.getBound(20, 1, user) + this.getLuckFactor(user) > 10;
+                        boolean luck = this.getBound(20, 1, user) + WeaponUtil.getLuckFactor(user) > 10;
                         LivingEntity living = luck ? target : user;
                         living.getArmorItems().iterator().forEachRemaining(itemStack -> {
                             boolean bl = this.random.nextBoolean();
@@ -106,7 +110,17 @@ public class WitheredWabbajackProjectile extends WitherSkullEntity {
     public void tick() {
         super.tick();
         if (this.age > 100) {
+            this.resetListLuckFactors();
             this.discard();
+        }
+    }
+
+    private void resetListLuckFactors() {
+        for (LuckChosenObject<?> luckChosenEntity : ENTITIES) {
+            luckChosenEntity.setLuckFactor(10);
+        }
+        for (LuckChosenObject<?> luckChosenEntity : COLLISIONS) {
+            luckChosenEntity.setLuckFactor(10);
         }
     }
 
@@ -125,11 +139,12 @@ public class WitheredWabbajackProjectile extends WitherSkullEntity {
         if (this.getOwner() != null && this.getOwner() instanceof LivingEntity) {
             this.randomCollisionEffect((LivingEntity)this.getOwner());
         }
+        this.resetListLuckFactors();
         this.discard();
     }
 
     private void randomCollisionEffect(LivingEntity user) {
-        int power = this.getBound(10 , 1, user) + this.getLuckFactor(user);
+        int power = this.getBound(10 , 1, user) + WeaponUtil.getLuckFactor(user);
         boolean unluckyAf = this.getBound(100 , 10, user) == 1;
         if (unluckyAf) {
             boolean isWarden = this.random.nextBoolean();
@@ -223,175 +238,73 @@ public class WitheredWabbajackProjectile extends WitherSkullEntity {
     }
 
     private int getBound(int bound, int luckModifier, LivingEntity user) {
-        int b = bound + this.getLuckFactor(user) * luckModifier;
+        int b = bound + WeaponUtil.getLuckFactor(user) * luckModifier;
         return b > 0 ? this.random.nextInt(b) : 1;
     }
 
     private CollisionEffect getCollisionEffectType(LivingEntity user) {
-        return (CollisionEffect) this.randomFromList(user, this.getTriggerList(), false);
+        return WeaponUtil.getRandomlyChosenObject(user, COLLISIONS, false);
     }
 
     private EntityType<?> getEntityType(LivingEntity user) {
-        return (EntityType<?>) this.randomFromList(user, this.getEntityList(), false);
-    }
-
-    private Object[][] getTriggerList() {
-        return new Object[][]{
-            {CollisionEffect.LIGHTNING, LuckType.GOOD},
-            {CollisionEffect.RANDOM_ENTITY, LuckType.BAD},
-            {CollisionEffect.SPECIFIC_ENTITY, LuckType.BAD},
-            {CollisionEffect.BATS, LuckType.BAD},
-            {CollisionEffect.PARTICLES, LuckType.BAD},
-            {CollisionEffect.DARKNESS, LuckType.BAD},
-            {CollisionEffect.EXPLOSION, LuckType.GOOD},
-        };
-    }
-
-    enum CollisionEffect {
-        LIGHTNING, RANDOM_ENTITY, SPECIFIC_ENTITY, BATS, PARTICLES, DARKNESS, EXPLOSION, CURSE
+        return WeaponUtil.getRandomlyChosenObject(user, ENTITIES, false);
     }
 
     private StatusEffect getRandomEffect(boolean flipLuckTypes) {
         if (this.getOwner() instanceof LivingEntity) {
-            return (StatusEffect) this.randomFromList((LivingEntity)this.getOwner(), this.getEffectObjectList(), flipLuckTypes);
+            return WeaponUtil.getRandomlyChosenObject((LivingEntity)this.getOwner(), this.getEffectList(), flipLuckTypes);
         } else {
             //Glowing as default incase nothing works
             return StatusEffects.GLOWING;
         }
     }
 
-    /**
-     * The new way to get a random object while including the status effect Luck found through
-     * {@link #getLuckFactor(LivingEntity)} as a factor. As long as the given array contains exactly an
-     * element with a given {@link LuckType}, an exception won't be thrown and the game won't crash :)
-     * <p></p>
-     * EDIT: Cool message, past me! But this sucks! Message to future me: rework it one day when you feel like it :)
-     */
-    private Object randomFromList(LivingEntity user, Object[][] arr, boolean flipLuckTypes) {
-        ArrayList<ArrayList<Object>> objectList = new ArrayList<>();
-        for (Object[] objects : arr) {
-            ArrayList<Object> list = new ArrayList<>();
-            list.add(objects[0]);
-            list.add(objects[1]);
-
-            int luckFactor;
-            if (!flipLuckTypes) {
-                switch ((LuckType) list.get(1)) {
-                    case BAD -> {
-                        luckFactor = (10 - this.getLuckFactor(user));
-                        list.add(luckFactor);
-                    }
-                    case GOOD -> {
-                        luckFactor = (10 + this.getLuckFactor(user));
-                        list.add(luckFactor);
-                    }
-                    default -> {
-                        luckFactor = 10;
-                        list.add(luckFactor);
-                    }
-                }
-            } else {
-                switch ((LuckType) list.get(1)) {
-                    case GOOD -> {
-                        luckFactor = (10 - this.getLuckFactor(user));
-                        list.add(luckFactor);
-                    }
-                    case BAD -> {
-                        luckFactor = (10 + this.getLuckFactor(user));
-                        list.add(luckFactor);
-                    }
-                    default -> {
-                        luckFactor = 10;
-                        list.add(luckFactor);
-                    }
-                }
-            }
-
-            if (!(luckFactor <= 0)) {
-                objectList.add(list);
-            }
-        }
-
-        int totalChance = 0;
-        for (ArrayList<Object> objects : objectList) {
-            totalChance += (int) objects.get(2);
-        }
-
-        int random = new Random().nextInt(totalChance);
-        int counter = 0;
-        int under = 0;
-        Object chosenObject = null;
-        for (int i = 0; i < objectList.size(); i++) {
-            counter += (int) objectList.get(i).get(2);
-            if (i > 0) {
-                under += (int) objectList.get(i - 1).get(2);
-            } 
-            if (random < counter && random >= under) {
-                chosenObject = objectList.get(i).get(0);
-            }
-        }
-        return chosenObject;
-    }
-
-    private HashMap<StatusEffect, LuckType> getEffectMap() {
-        HashMap<StatusEffect, LuckType> map = new HashMap<>();
+    private List<LuckChosenObject<StatusEffect>> getEffectList() {
+        List<LuckChosenObject<StatusEffect>> list = new ArrayList<>();
         for (StatusEffect effect : Registries.STATUS_EFFECT) {
             if (effect.getCategory().equals(StatusEffectCategory.HARMFUL)) {
-                map.put(effect, LuckType.BAD);
+                list.add(new LuckChosenObject<>(effect, WeaponUtil.LuckType.BAD));
             } else if (effect.getCategory().equals(StatusEffectCategory.BENEFICIAL)) {
-                map.put(effect, LuckType.GOOD);
+                list.add(new LuckChosenObject<>(effect, WeaponUtil.LuckType.GOOD));
             } else {
-                map.put(effect, LuckType.NEUTRAL);
+                list.add(new LuckChosenObject<>(effect, WeaponUtil.LuckType.NEUTRAL));
             }
         }
-        return map;
+        return list;
     }
 
-    /**
-     * I'm not going to bother reworking this class today so this will do :)
-     */
-    private Object[][] getEffectObjectList() {
-        HashMap<StatusEffect, LuckType> map = this.getEffectMap();
-        int len = map.size();
-        Object[][] arr = new Object[len][2];
-        int i = 0;
-        for (StatusEffect effect : map.keySet()) {
-            arr[i][0] = effect;
-            arr[i][1] = map.get(effect);
-            i++;
-        }
-        return arr;
+    static {
+        ENTITIES.add(new LuckChosenObject<>(EntityType.CREEPER, WeaponUtil.LuckType.BAD));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.ZOMBIE, WeaponUtil.LuckType.BAD));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.ENDERMITE, WeaponUtil.LuckType.BAD));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.BEE, WeaponUtil.LuckType.BAD));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.PUFFERFISH, WeaponUtil.LuckType.BAD));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.VEX, WeaponUtil.LuckType.BAD));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.SKELETON, WeaponUtil.LuckType.BAD));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.DROWNED, WeaponUtil.LuckType.BAD));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.GUARDIAN, WeaponUtil.LuckType.BAD));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.SALMON, WeaponUtil.LuckType.NEUTRAL));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.COW, WeaponUtil.LuckType.NEUTRAL));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.COD, WeaponUtil.LuckType.NEUTRAL));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.TROPICAL_FISH, WeaponUtil.LuckType.NEUTRAL));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.WANDERING_TRADER, WeaponUtil.LuckType.NEUTRAL));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.CHICKEN, WeaponUtil.LuckType.NEUTRAL));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.SHEEP, WeaponUtil.LuckType.NEUTRAL));
+        ENTITIES.add(new LuckChosenObject<>(EntityType.EXPERIENCE_ORB, WeaponUtil.LuckType.GOOD));
     }
 
-    private Object[][] getEntityList() {
-        return new Object[][]{
-            {EntityType.CREEPER, LuckType.BAD},
-            {EntityType.ZOMBIE, LuckType.BAD},
-            {EntityType.COW, LuckType.NEUTRAL},
-            {EntityType.COD, LuckType.NEUTRAL},
-            {EntityType.ENDERMITE, LuckType.BAD},
-            {EntityType.BEE, LuckType.BAD},
-            {EntityType.EXPERIENCE_ORB, LuckType.GOOD},
-            {EntityType.SALMON, LuckType.NEUTRAL},
-            {EntityType.PUFFERFISH, LuckType.BAD},
-            {EntityType.VEX, LuckType.BAD},
-            {EntityType.SKELETON, LuckType.BAD},
-            {EntityType.TROPICAL_FISH, LuckType.NEUTRAL},
-            {EntityType.WANDERING_TRADER, LuckType.NEUTRAL},
-            {EntityType.CHICKEN, LuckType.NEUTRAL},
-            {EntityType.SHEEP, LuckType.NEUTRAL},
-            {EntityType.DROWNED, LuckType.BAD},
-            {EntityType.GUARDIAN, LuckType.BAD},
-        };
+    static {
+        COLLISIONS.add(new LuckChosenObject<>(CollisionEffect.RANDOM_ENTITY, WeaponUtil.LuckType.BAD));
+        COLLISIONS.add(new LuckChosenObject<>(CollisionEffect.SPECIFIC_ENTITY, WeaponUtil.LuckType.BAD));
+        COLLISIONS.add(new LuckChosenObject<>(CollisionEffect.BATS, WeaponUtil.LuckType.BAD));
+        COLLISIONS.add(new LuckChosenObject<>(CollisionEffect.PARTICLES, WeaponUtil.LuckType.BAD));
+        COLLISIONS.add(new LuckChosenObject<>(CollisionEffect.DARKNESS, WeaponUtil.LuckType.BAD));
+        COLLISIONS.add(new LuckChosenObject<>(CollisionEffect.CURSE, WeaponUtil.LuckType.BAD));
+        COLLISIONS.add(new LuckChosenObject<>(CollisionEffect.LIGHTNING, WeaponUtil.LuckType.GOOD));
+        COLLISIONS.add(new LuckChosenObject<>(CollisionEffect.EXPLOSION, WeaponUtil.LuckType.GOOD));
     }
 
-    private int getLuckFactor(LivingEntity entity) {
-        if (entity.hasStatusEffect(StatusEffects.LUCK)) {
-            return entity.getStatusEffect(StatusEffects.LUCK).getAmplifier() * 2 + 2;
-        } else if (entity.hasStatusEffect(StatusEffects.UNLUCK)) {
-            return - entity.getStatusEffect(StatusEffects.UNLUCK).getAmplifier() * 2 + 2;
-        } else {
-            return 0;
-        }
+    enum CollisionEffect {
+        LIGHTNING, RANDOM_ENTITY, SPECIFIC_ENTITY, BATS, PARTICLES, DARKNESS, EXPLOSION, CURSE
     }
 }